fix tests with new api add test for runtime for local
